Don’t activate inside child frames, change stylesheet override behavior
Dynamically-added iframes would trigger Nightlight, which was annoying if the user toggled Nightlight off for the page. This change means that frames won’t be styled, which needs to be fixed.

Sometimes the original stylesheet took precedence over Nightlight’s new stylesheet, resulting in elements not being darkened; the fix is to disable the original stylesheet, which requires changing logic in makeStyle.